🐛 Profile button now edits profile
- refactor extract editperson and PersonEditor
